even less engine responsibilities
- decided on using dedicated file for fastforwarding
- engines no longer open/read/write tag-, blacklist-, or
  fastforward files
- engine.prepare() is called only when the above required files exist
    - custom files needed for a specific engine must still be manually
      handled in engine.prepare()

basically, the idea from this point on is to have the engines
responsible for as little as possible.  sadly, the engine interface is
growing quite a bit, but the majority of engine methods simply return
strings that are used for populating required files, so at least this is
not difficult for a dev to add.

next is to experiment a bit with dict.update() to hopefully simplify
error cases by setting engine.state = False